\name{dataGetDetailStore}
\alias{dataGetDetailStore}
\title{Baixa tabelas com informações detalhadas das lojas}
\usage{
dataGetDetailStore()
}



\value{
Esta função retorna uma matriz de detalhes das lojas. 

}


\examples{
#Lojas
m=dataGetDetailStore()

}

\author{
  Coded by Henrique Helfer Hoeltgebaum and Gabriel Vasconcelos, last updated on 01-11-2016.
  }
