% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataProcess.R
\name{gdp}
\alias{gdp}
\title{Title}
\usage{
gdp(
  Kvalue,
  Tvalue,
  paramDelta,
  paramOmega,
  paramAlpha,
  paramBeta,
  vectorSizeIntensity,
  init,
  xRates
)
}
\arguments{
\item{xRates}{}
}
\description{
Title
}
