% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodAutoregressiveRepeat.R
\name{likelihoodAutoregressiveRepeat}
\alias{likelihoodAutoregressiveRepeat}
\title{Title}
\usage{
likelihoodAutoregressiveRepeat(
  listOfData,
  vectorOfParameters,
  yvar,
  ysize,
  Xvar,
  skipto
)
}
\arguments{
\item{skipto}{}
}
\description{
Title
}
