% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_repos.R
\name{repeat_repos}
\alias{repeat_repos}
\title{Title}
\usage{
repeat_repos(
  dataList,
  yvar,
  Xvar,
  ysize,
  model_order,
  skipto = NULL,
  repeat_opt = repeat_opt
)
}
\arguments{
\item{repeat_opt}{}
}
\description{
Title
}
