% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_repos.R
\name{simple_repos}
\alias{simple_repos}
\title{Title}
\usage{
simple_repos(
  data,
  yvar,
  Xvar,
  ysize,
  model_order,
  skipto = NULL,
  repeat_opt = repeat_opt,
  stepwise = TRUE
)
}
\arguments{
\item{repeat_opt}{}
}
\description{
Title
}
