% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_repos.R
\name{simulate_repos}
\alias{simulate_repos}
\title{Title}
\usage{
simulate_repos(
  Tvalue,
  paramDelta,
  paramOmega,
  paramAlpha,
  paramBeta,
  sizeIntensity,
  init,
  xRates,
  model_order
)
}
\arguments{
\item{model_order}{}
}
\description{
Title
}
