% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdErrorAutoregressive.R
\name{stdErrorAutoregressive}
\alias{stdErrorAutoregressive}
\title{Title}
\usage{
stdErrorAutoregressive(
  listOfData,
  vectorOfParameters,
  yvar,
  Xvar,
  ysize,
  model_order,
  skipto
)
}
\arguments{
\item{skipto}{}
}
\description{
Title
}
