% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdErrorAutoregressiveRepeat.R
\name{stdErrorAutoregressiveRepeat}
\alias{stdErrorAutoregressiveRepeat}
\title{Title}
\usage{
stdErrorAutoregressiveRepeat(
  listOfData,
  vectorOfParameters,
  yvar,
  ysize,
  Xvar,
  skipto
)
}
\arguments{
\item{skipto}{}
}
\description{
Title
}
