% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_repos.R
\name{stepwise_repos}
\alias{stepwise_repos}
\title{Title}
\usage{
stepwise_repos(
  data,
  yvar,
  Xvar,
  ysize,
  model_order,
  skipto = NULL,
  repeat_opt = NULL,
  stepwise = stepwise
)
}
\arguments{
\item{stepwise}{}
}
\description{
Title
}
