% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos}
\alias{repos}
\title{Title}
\usage{
repos(
  data,
  y,
  exp_var,
  size,
  model_order = 2,
  skipto = 10,
  repeat_opt = 20,
  stepwise = FALSE
)
}
\arguments{
\item{repeat_opt}{}
}
\description{
Title
}
